Document status fields more clearly.
It is a bit confusing because the `when` field is really a staleness metric but is inside the variants, it applies to the status as a whole not something specific to the variant. Add documentation to be a bit more clear. It isn't worth breaking compatibility to fix, but docs are nice.