Add names to tunnels.
I went back and forth on the representation of this. For example making it an always-present string that may be empty and empty means not set. This is nice because it removes one possible value from the type. However I decided to go with `Option` so that handling the basic flow is natural even if the ideal UX would do things like `tunnel.name.filter(|s| s.is_all_whitespace())`. The API will treat `Some("")` and `None` identically on input to effectively remove this case even if it is possible in the type definitions. Furthermore these cleanups should really be done on the client if they are done at all.

Fixes: https://linear.app/soveng/issue/OBS-1462/add-names-to-tunnels