Update the `gen_id` example to create full-entropy account numbers.
The range was originally limited to i64 to making working with the unchecked account number simpler in some situations, however it isn't worth having this command generated numbers with less entropy without notice.

The account number generator in the macOS client and website have always generated numbers with the full entropy. So this only affects numbers manually generate during testing.